Added the new DebugWithAnalysis configuration that runs clang-tidy after build. Sorted out a few of the warnings it produced.

+++ src/application/sdl_application.h
@@ -33,6 +33,6 @@ private:
 	int m_audioDeviceId;
 	bool m_keepWindowOpen;
 
-	float m_timeLastFrame;
+	int m_timeLastFrame;
 	int m_barWidth;
 };
